Add "signed-in" and "edit access" checks

+++ routes.py
@@ -1,6 +1,6 @@
 from sqlalchemy import true
-from app import app
 from flask import render_template, redirect, request, session, abort
+from app import app
 import users, messages
 
 
@@ -12,6 +12,9 @@ def index():
 
 @app.route("/topic/<int:topic_id>")  # 7.4
 def topic(topic_id):
+    if users.has_access(topic_id) is None: # 22.4
+        return render_template("error.html", message="Käyttäjällä ei oikeutta sivuun")
+
     list = messages.get_threads(topic_id)
     name = messages.get_topic_name(topic_id)
     return render_template("topic.html", threads=list, topic_id=topic_id, topic_name=name)
@@ -19,6 +22,9 @@ def topic(topic_id):
 
 @app.route("/topic/<int:topic_id>/thread/<int:thread_id>")
 def thread(topic_id, thread_id):
+    if users.has_access(topic_id) is None: # 22.4
+        return render_template("error.html", message="Käyttäjällä ei oikeutta sivuun")
+
     list = messages.get_messages(thread_id)
     path = messages.get_path(topic_id, thread_id)
     return render_template("thread.html", messages=list, topic_id=topic_id, thread_id=thread_id, message_path=path)
@@ -55,7 +61,7 @@ def register():
         if len(username) < 1:
             return render_template("error.html", message="Käyttäjätunnus liian lyhyt")
         if len(username) > 50:
-            return render_template("error.html", message="Käyttäjätunnus liian pitkä")    
+            return render_template("error.html", message="Käyttäjätunnus liian pitkä")
         password1 = request.form["password1"]
         password2 = request.form["password2"]
 
@@ -68,7 +74,7 @@ def register():
         else:
             return render_template("error.html", message="Rekisteröinti ei onnistunut, valitse toinen käyttäjätunnus")
 
-@app.route("/create_new_topic") #11.4, 
+@app.route("/create_new_topic") #11.4,
 def create_new_topic():
     return render_template("new_topic_template.html")
 
@@ -91,15 +97,29 @@ def new_topic():
     return redirect("/")
 
 
-@app.route("/hide_topic/<int:topic_id>", methods=["GET"])   #20.4
-def hide_topic(topic_id):    
-    messages.hide_topic(topic_id)
-    return redirect("/") 
 
-@app.route("/hide_secret_topic/<int:topic_id>", methods=["GET"])   #20.4
-def hide_secret_topic(topic_id):    
+@app.route("/hide_topic/<int:topic_id>", methods=["GET"])   #20.4 + 22.4
+def hide_topic(topic_id):
+    if not session.get("admin", 0):   #if not admin, can't hide topics
+        abort(403)
+
+    if session["csrf_token"] != request.form["csrf_token"]:    # to take into acc. CSRF-vulnerability
+        abort(403)
+
+    messages.hide_topic(topic_id)
+    return redirect("/")
+
+
+@app.route("/hide_secret_topic/<int:topic_id>", methods=["GET"])   #20.4 +22.4
+def hide_secret_topic(topic_id):
+    if not session.get("admin", 0):   #if not admin, can't hide topics
+        abort(403)
+
+    if session["csrf_token"] != request.form["csrf_token"]:    # to take into acc. CSRF-vulnerability
+        abort(403)
+
     messages.hide_secret_topic(topic_id)
-    return redirect("/") 
+    return redirect("/")
 
 
 @app.route("/access/<int:topic_id>")    #13.4
@@ -116,11 +136,11 @@ def grant_topic_access():
         abort(403)
 
     topic_id = request.form["topic_id"]
-    username=request.form["username"]
+    username = request.form["username"]
 
     if len(username) < 1 or len(username) > 50:
             return render_template("error.html", message="Käyttäjätunnuksen pituus on 1-50 merkkiä")
-       
+
     user_id = users.get_user_id(username)
 
     if not user_id:
@@ -139,11 +159,14 @@ def new(topic_id):
 def new_thread():
     user_id = users.user_id()
 
+    if user_id == 0:   # 22.4 has to be signed-in
+        return render_template("error.html", message="Et ole kirjautunut sisään.")
+
     if session["csrf_token"] != request.form["csrf_token"]:    # to take into acc. CSRF-vulnerability
         abort(403)
 
-    thread_name = request.form["thread_name"]   
-    first_message = request.form["message"] 
+    thread_name = request.form["thread_name"]
+    first_message = request.form["message"]
     topic_id = request.form["topic_id"]
 
     if len(thread_name) < 1 or len(thread_name) > 100:
@@ -166,11 +189,19 @@ def edit_thread(topic_id, thread_id):
 
 @app.route("/edit_thread_header", methods=["POST"])    #18.4
 def edit_thread_header():
+    user_id = users.user_id()
+
+    if user_id == 0:   # 22.4 has to be signed-in
+        return render_template("error.html", message="Et ole kirjautunut sisään.")
+
     if session["csrf_token"] != request.form["csrf_token"]:    # to take into acc. CSRF-vulnerability
         abort(403)
-    thread_name = request.form["thread_name"]  
-    thread_id = request.form["thread_id"]  
+    thread_name = request.form["thread_name"]
+    thread_id = request.form["thread_id"]
     topic_id = request.form["topic_id"]
+
+    if users.thread_edit_access(thread_id) is None: # 23.4 only if own or admin
+        return render_template("error.html", message="Käyttäjällä ei oikeutta muokata ketjua.")
 
     if len(thread_name) < 1 or len(thread_name) > 100:
         return render_template("error.html", message="Keskustelun otsikon tulee olla 1-100 merkkiä.")
@@ -183,36 +214,46 @@ def edit_thread_header():
 @app.route("/delete_thread/<int:thread_id>", methods=["GET"])   #18.4
 def delete_thread(thread_id):
     #if session["csrf_token"] != request.form["csrf_token"]:    # to take into acc. CSRF-vulnerability
-    #    abort(403)   #check if needed??  
+    #    abort(403)   #check if needed??
+    user_id = users.user_id()
+
+    if user_id == 0:   # 22.4 has to be signed-in
+        return render_template("error.html", message="Et ole kirjautunut sisään.")
+
+    if users.thread_edit_access(thread_id) is None: # 23.4 only if own or admin
+        return render_template("error.html", message="Käyttäjällä ei oikeutta muokata ketjua.")
+
     topic_id = messages.delete_this_thread(thread_id)[0]
 
     return redirect("/topic/" + str(topic_id))
 
 
 
-@app.route("/message/<int:topic_id>/<int:thread_id>")   
+@app.route("/message/<int:topic_id>/<int:thread_id>")
 def message(topic_id, thread_id):     #fixed 9.4,
-    return render_template("message.html", topic_id=topic_id, thread_id=thread_id)   #  add topic 4.4 
+    return render_template("message.html", topic_id=topic_id, thread_id=thread_id)   #  add topic 4.4
 
 
 @app.route("/new_message", methods=["POST"])   # modified 9.4
 def new_message():
-    #check if signed-in???
     user_id = users.user_id()
+
+    if user_id == 0:   # 22.4 has to be signed-in
+        return render_template("error.html", message="Et ole kirjautunut sisään.")
 
     if session["csrf_token"] != request.form["csrf_token"]:    # to take into acc. CSRF-vulnerability
         abort(403)
 
-    message = request.form["message"]   
-    topic_id = request.form["topic_id"]   
-    thread_id = request.form["thread_id"]    
+    message = request.form["message"]
+    topic_id = request.form["topic_id"]
+    thread_id = request.form["thread_id"]
 
     if len(message) > 5000:
         return render_template("error.html", message="Viesti on liian pitkä.") #
 
     if len(message) < 1:
-        return render_template("error.html", message="Viesti on liian lyhyt.") #    
-   
+        return render_template("error.html", message="Viesti on liian lyhyt.") #
+
     if messages.send(message, thread_id, user_id):
         #return redirect("/")   # or return redirect("/topic"+ topic_id)  ??
         return redirect("/topic/" + str(topic_id) + "/thread/" + str(thread_id))
@@ -222,24 +263,32 @@ def new_message():
 
 @app.route("/edit_message/<int:topic_id>/<int:thread_id>/<int:message_id>")   #19.4 & 21.4
 def edit_message(topic_id, thread_id, message_id): # include old content
-    message = messages.get_message_to_edit(message_id) 
+    message = messages.get_message_to_edit(message_id)
     return render_template("edit_message.html", topic_id=topic_id, thread_id=thread_id, message_id=message_id, message=message)
 
 
 @app.route("/edit_message_text", methods=["POST"])   #19.4
 def edit_message_text():
+    user_id = users.user_id()
+
+    if user_id == 0:   # 22.4 has to be signed-in
+        return render_template("error.html", message="Et ole kirjautunut sisään.")
+
     if session["csrf_token"] != request.form["csrf_token"]:    # to take into acc. CSRF-vulnerability
         abort(403)
-    message = request.form["message"]  
-    topic_id = request.form["topic_id"]   
-    thread_id = request.form["thread_id"] 
-    message_id = request.form["message_id"]  
+    message = request.form["message"]
+    topic_id = request.form["topic_id"]
+    thread_id = request.form["thread_id"]
+    message_id = request.form["message_id"]
+
+    if users.message_edit_access(message_id) is None: # 23.4 only if own msg or admin
+        return render_template("error.html", message="Käyttäjällä ei oikeutta muokata viestiä.")
 
     if len(message) > 5000:
         return render_template("error.html", message="Viesti on liian pitkä.") #
 
     if len(message) < 1:
-        return render_template("error.html", message="Viesti on liian lyhyt.") # 
+        return render_template("error.html", message="Viesti on liian lyhyt.") #
 
     messages.edit_message(message, message_id)
     return redirect("/topic/" + str(topic_id) + "/thread/" + str(thread_id))
@@ -248,7 +297,15 @@ def edit_message_text():
 @app.route("/delete_message/<int:topic_id>/<int:thread_id>/<int:message_id>", methods=["GET"])   #19.4
 def delete_message(topic_id, thread_id, message_id):
     #if session["csrf_token"] != request.form["csrf_token"]:    # to take into acc. CSRF-vulnerability
-    #    abort(403)   #check if needed  
+    #    abort(403)   #check if needed
+    user_id = users.user_id()
+
+    if user_id == 0:   # 22.4 has to be signed-in
+        return render_template("error.html", message="Et ole kirjautunut sisään.")
+
+    if users.message_edit_access(message_id) is None: # 23.4 only if own msg or admin
+        return render_template("error.html", message="Käyttäjällä ei oikeutta poistaa viestiä.")
+
     thread_id = messages.delete_this_message(message_id)[0]
 
     return redirect("/topic/" + str(topic_id) + "/thread/" + str(thread_id))
@@ -256,16 +313,19 @@ def delete_message(topic_id, thread_id, message_id):
 
 @app.route("/search")    #16.4
 def search():
+    user_id = users.user_id()
+    if user_id == 0:   # 22.4 has to be signed-in
+        return render_template("error.html", message="Voit hakea viestejä, kun olet kirjautunut sisään.")
     return render_template("search.html")
 
 @app.route("/search_result", methods=["GET"])    #16.4
 def search_result():
     query = request.args["query"]
     if len(query) < 1 or len(query) > 50:
-        return render_template("error.html", message = "Haettavan sanan pituus voi olla 1-50 merkkiä.")
+        return render_template("error.html", message="Haettavan sanan pituus voi olla 1-50 merkkiä.")
     result_messages = messages.search_message(query)
 
     if result_messages is None or len(result_messages) == 0:
-        return render_template("error.html", message = "Hakusanalla ei löydy viestejä.")
+        return render_template("error.html", message="Hakusanalla ei löydy viestejä.")
 
-    return render_template("search_result.html", messages = result_messages)
+    return render_template("search_result.html", messages=result_messages)
